polaire/requirements.txt
django
djangorestframework
django-cors-headers
django-filter
djangorestframework-jwt
psycopg2-binary